Configurable root node via props
Using a Picker inside a modal window, there are issues with onBlur events.The calendarContainer is in the document.body and picks a day fires the onBlur's modal window because the calendarContainer is not inside the modal div. This PR adds configuration for the root node to anchor the calendar's container. 

For example:
&lt;Picker rootNode={React.findDOMNode(this.refs.divContainer)} adjustOrientOnCalendarOverflow={true} ... /&gt;